fix: dialogs fucking around :c
